install: make it impossible for root to leave files behind
If root leaves a file in a directory under /var/lib/wazo-plugind/plugins while
installing a package, it will not be possible to remove that directory without
being root. Making the API unusable or requiring most operations to be done as
root.

Using chattr +i makes the whole plugins directory immutable for the time of the
install.