remove the /var/lib/wazo-plugind/plugins directory
This directory used to be a storage for installed plugins that could be used to
query and uninstall plugins. Since we are now using debian packages this
directory is not necessary anymore.

Removing it also remove some case of race condition that could happen while
installing the same plugin with 2 simultaneous requests.