db: fix StrictVersion usage
If the value is an Integer, then a TypeError is raised. If we catch
TypeError but do not convert the value to string AND there are string
version in others plugin, then python can't sort str() with int().

If the value is None, then:
StrictVersion(None)  # raise AttributeError

value = StrictVersion(None)  # raise nothing
value  # raise AttributeError