worker: add a lock around the put and get of the worker queues
When many simultaneous installation used the worker each request could put
a task on the command queue and call get blocking until a task completed.

The problem is that a thread could receive the result of another task.

Protecting the two opetations with a lock will not allow a task to be started
by a thread until the worker has completed its previous task. This does not
change the time spent for the overall process since the worker have always been
sequential.