refactor: replace args by options
at the begining args in the plugin post were prety much anything. Now they are "typed"
depending of the method used for installation and theres also a additionnal params
that can be added using the query string.

I renamed it to install_options to reflect the body of the post and avoid the temptation
to add other fields to it that are not related to the used backend.